Refactor to remove cyclic dependencies
