+++ package.json
@@ -28,6 +28,7 @@
     "rollup-plugin-typescript2": "0.13.0",
     "rollup-plugin-uglify": "~3.0.0",
     "tslint": "~5.6.0",
+    "tslint-no-circular-imports": "^0.5.0",
     "typescript": "2.8.3",
     "yalc": "~1.0.0-pre.21"
   },
