+++ src/callbacks_test.ts
@@ -15,16 +15,11 @@
 // tslint:disable:max-line-length
 import {scalar} from '@tensorflow/tfjs-core';
 
-import {BaseLogger, CallbackList, History, resolveScalarsInLogs, UnresolvedLogs, disposeTensorsInLogs} from './callbacks';
-import {Model} from './engine/training';
+import {disposeTensorsInLogs, resolveScalarsInLogs, UnresolvedLogs} from './engine/logs';
+import {BaseLogger, History} from './engine/training';
 import {describeMathCPUAndGPU} from './utils/test_utils';
-// tslint:enable:max-line-length
 
-class MockModel extends Model {
-  constructor(name: string) {
-    super({inputs: [], outputs: [], name});
-  }
-}
+// tslint:enable:max-line-length
 
 describe('BaseLogger Callback', () => {
   it('Records and averages losses in an epoch', async done => {
@@ -93,75 +88,6 @@ describe('History Callback', () => {
   });
 });
 
-describe('CallbackList', () => {
-  it('Constructor with empty arg', async done => {
-    const callbackList = new CallbackList();
-    await callbackList.onTrainBegin();
-    await callbackList.onTrainEnd();
-    done();
-  });
-  it('Constructor and setParams with array of callbacks', () => {
-    const history1 = new History();
-    const history2 = new History();
-    const callbackList = new CallbackList([history1, history2]);
-    const params = {'verbose': 3};
-    callbackList.setParams(params);
-    expect(history1.params).toEqual(params);
-    expect(history2.params).toEqual(params);
-  });
-  it('Constructor and setModel with array of callbacks', () => {
-    const history1 = new History();
-    const history2 = new History();
-    const callbackList = new CallbackList([history1, history2]);
-    const model = new MockModel('MockModelA');
-    callbackList.setModel(model);
-    expect(history1.model).toEqual(model);
-    expect(history2.model).toEqual(model);
-  });
-  it('onTrainBegin', async done => {
-    const history1 = new History();
-    const history2 = new History();
-    const callbackList = new CallbackList([history1, history2]);
-    await callbackList.onTrainBegin();
-    expect(history1.epoch).toEqual([]);
-    expect(history1.history).toEqual({});
-    expect(history2.epoch).toEqual([]);
-    expect(history2.history).toEqual({});
-    done();
-  });
-  it('onEpochEnd', async done => {
-    const history1 = new History();
-    const history2 = new History();
-    const callbackList = new CallbackList([history1, history2]);
-    await callbackList.onTrainBegin();
-    await callbackList.onEpochEnd(100, {'val_loss': 10, 'val_accuracy': 0.1});
-    expect(history1.epoch).toEqual([100]);
-    expect(history1.history).toEqual({'val_loss': [10], 'val_accuracy': [0.1]});
-    expect(history2.epoch).toEqual([100]);
-    expect(history2.history).toEqual({'val_loss': [10], 'val_accuracy': [0.1]});
-    await callbackList.onEpochEnd(101, {'val_loss': 9.5, 'val_accuracy': 0.2});
-    expect(history1.epoch).toEqual([100, 101]);
-    expect(history1.history)
-        .toEqual({'val_loss': [10, 9.5], 'val_accuracy': [0.1, 0.2]});
-    expect(history2.epoch).toEqual([100, 101]);
-    expect(history2.history)
-        .toEqual({'val_loss': [10, 9.5], 'val_accuracy': [0.1, 0.2]});
-    done();
-  });
-  it('append', async done => {
-    const history1 = new History();
-    const history2 = new History();
-    const callbackList = new CallbackList([history1]);
-    await callbackList.onTrainBegin();
-    expect(history1.epoch).toEqual([]);
-    expect(history1.history).toEqual({});
-    await callbackList.append(history2);
-    await callbackList.onTrainBegin();
-    expect(history2.epoch).toEqual([]);
-    expect(history2.history).toEqual({});
-    done();
-  });
-});
 
 describeMathCPUAndGPU('resolveScalarsInLogs', () => {
   it('Resolve mixed numbers and scalars', async done => {
@@ -210,5 +136,4 @@ describeMathCPUAndGPU('disposeTensorsInLogs', () => {
     // tslint:disable-next-line:no-any
     expect((logs['d'] as any).isDisposed).toEqual(true);
   });
-
 });
