+++ src/constraints.ts
@@ -14,7 +14,9 @@
 import * as tfc from '@tensorflow/tfjs-core';
 import {doc, serialization, Tensor, tidy} from '@tensorflow/tfjs-core';
 
-import * as K from './backend/tfjs_backend';
+import {epsilon} from './backend/common';
+import {getScalar} from './engine/unique_id';
+// import * as K from './backend/tfjs_backend';
 import {deserializeKerasObject, serializeKerasObject} from './utils/generic_utils';
 // tslint:enable:max-line-length
 
@@ -22,7 +24,7 @@ import {deserializeKerasObject, serializeKerasObject} from './utils/generic_util
  * Helper function used by many of the Constraints to find the L2Norms.
  */
 function calcL2Norms(w: Tensor, axis: number): Tensor {
-  return tidy(() => tfc.sqrt(tfc.sum(K.square(w), axis, true)));
+  return tidy(() => tfc.sqrt(tfc.sum(tfc.mulStrict(w, w), axis, true)));
 }
 
 /**
@@ -88,9 +90,7 @@ export class MaxNorm extends Constraint {
     return tidy(() => {
       const norms = calcL2Norms(w, this.axis);
       const desired = tfc.clipByValue(norms, 0, this.maxValue);
-      return tfc.mul(
-          w,
-          tfc.div(desired, K.scalarPlusArray(K.getScalar(K.epsilon()), norms)));
+      return tfc.mul(w, tfc.div(desired, tfc.add(getScalar(epsilon()), norms)));
     });
   }
 
@@ -133,9 +133,7 @@ export class UnitNorm extends Constraint {
   apply(w: Tensor): Tensor {
     return tidy(
         () => tfc.div(
-            w,
-            K.scalarPlusArray(
-                K.getScalar(K.epsilon()), calcL2Norms(w, this.axis))));
+            w, tfc.add(getScalar(epsilon()), calcL2Norms(w, this.axis))));
   }
 
   getConfig(): serialization.ConfigDict {
@@ -215,13 +213,11 @@ export class MinMaxNorm extends Constraint {
     return tidy(() => {
       const norms = calcL2Norms(w, this.axis);
       const desired = tfc.add(
-          K.scalarTimesArray(
-              K.getScalar(this.rate),
+          tfc.mul(
+              getScalar(this.rate),
               tfc.clipByValue(norms, this.minValue, this.maxValue)),
-          K.scalarTimesArray(K.getScalar(1.0 - this.rate), norms));
-      return tfc.mul(
-          w,
-          tfc.div(desired, K.scalarPlusArray(K.getScalar(K.epsilon()), norms)));
+          tfc.mul(getScalar(1.0 - this.rate), norms));
+      return tfc.mul(w, tfc.div(desired, tfc.add(getScalar(epsilon()), norms)));
     });
   }
