+++ src/engine/training.ts
@@ -12,22 +12,28 @@
 
 // tslint:disable:max-line-length
 import * as tfc from '@tensorflow/tfjs-core';
-import {doc, io, ModelPredictConfig, Optimizer, Scalar, serialization, Tensor, Tensor1D, tensor1d, util} from '@tensorflow/tfjs-core';
+import {add, div, doc, io, keep, ModelPredictConfig, mul, Optimizer, Scalar, serialization, Tensor, Tensor1D, tensor1d, tidy, util} from '@tensorflow/tfjs-core';
 
 import * as K from '../backend/tfjs_backend';
-import {BaseLogger, Callback, CallbackList, CustomCallbackConfig, disposeTensorsInLogs, History, standardizeCallbacks, UnresolvedLogs} from '../callbacks';
+// import {BaseLogger, CustomCallbackConfig, disposeTensorsInLogs, History,
+// standardizeCallbacks} from '../callbacks';
+import {nameScope} from '../common';
 import {NotImplementedError, RuntimeError, ValueError} from '../errors';
 import * as losses from '../losses';
 import * as Metrics from '../metrics';
 import * as optimizers from '../optimizers';
-import {LossOrMetricFn, NamedTensorMap, Shape, SymbolicTensor} from '../types';
+import {LossOrMetricFn, NamedTensorMap, Shape} from '../types';
+import * as generic_utils from '../utils/generic_utils';
 import {count, pyListRepeat, singletonOrArray, unique} from '../utils/generic_utils';
 import {printSummary} from '../utils/layer_utils';
 import {range} from '../utils/math_utils';
 import {LayerVariable} from '../variables';
 
 import {execute, FeedDict} from './executor';
-import {Container, ContainerConfig} from './topology';
+import {disposeTensorsInLogs, Logs, resolveScalarsInLogs, UnresolvedLogs} from './logs';
+import {Container, ContainerConfig, SymbolicTensor} from './topology';
+import {getScalar} from './unique_id';
+
 // tslint:enable:max-line-length
 
 /**
@@ -794,7 +800,7 @@ export class Model extends Container implements tfc.InferenceModel {
     // Porting Note: In PyKeras, metrics_tensors are symbolic tensor objects.
     //   Here, metricsTensors are TypeScript functions. This difference is due
     //   to the difference in symbolic/imperative property of the backends.
-    K.nameScope('loss', () => {
+    nameScope('loss', () => {
       for (let i = 0; i < this.outputs.length; ++i) {
         if (skipTargetIndices.indexOf(i) !== -1) {
           continue;
@@ -828,7 +834,7 @@ export class Model extends Container implements tfc.InferenceModel {
           this.metricsTensors.push([metricTensor, outputIndex]);
         };
 
-    K.nameScope('metric', () => {
+    nameScope('metric', () => {
       for (let i = 0; i < this.outputs.length; ++i) {
         if (skipTargetIndices.indexOf(i) !== -1) {
           continue;
@@ -893,7 +899,7 @@ export class Model extends Container implements tfc.InferenceModel {
 
             // TODO(cais): Add weighting and masking to metricResult.
             let metricResult: LossOrMetricFn;
-            K.nameScope(metricName, () => {
+            nameScope(metricName, () => {
               metricResult = weightedMetricFn;
             });
             appendMetric(i, metricName, metricResult);
@@ -1489,20 +1495,20 @@ export class Model extends Container implements tfc.InferenceModel {
           const batchOuts = f(insBatch);
           if (batchIndex === 0) {
             for (let i = 0; i < batchOuts.length; ++i) {
-              outs.push(K.getScalar(0));
+              outs.push(getScalar(0));
             }
           }
           for (let i = 0; i < batchOuts.length; ++i) {
             const batchOut = batchOuts[i];
-            outs[i] = tfc.add(
-                          outs[i],
-                          K.scalarTimesArray(
-                              K.getScalar(batchEnd - batchStart), batchOut)) as
+            outs[i] =
+                tfc.add(
+                    outs[i],
+                    tfc.mul(getScalar(batchEnd - batchStart), batchOut)) as
                 Scalar;
           }
         }
         for (let i = 0; i < outs.length; ++i) {
-          outs[i] = tfc.div(outs[i], K.getScalar(numSamples)) as Scalar;
+          outs[i] = tfc.div(outs[i], getScalar(numSamples)) as Scalar;
         }
       }
       return outs;
@@ -1936,3 +1942,407 @@ export class Model extends Container implements tfc.InferenceModel {
 }
 
 serialization.SerializationMap.register(Model);
+
+
+
+export type Params = {
+  [key: string]: number|string|boolean|number[]|string[]|boolean[];
+};
+
+/**
+ * Abstract base class used to build new callbacks.
+ *
+ * The `logs` dictionary that callback methods take as argument will contain
+ * keys for quantities relevant to the current batch or epoch.
+ *
+ * Currently, the `.fit()` method of the `Sequential` model class
+ * will include the following quantities in the `logs` that
+ * it passes to its callbacks:
+ *
+ * onEpochEnd: Logs include `acc` and `loss`, and optionally include `valLoss`
+ *   (if validation is enabled in `fit`), and `valAcc` (if validation and
+ *   accuracy monitoring are enabled).
+ * onBatchBegin: Logs include `size`, the number of samples in the current
+ *   batch.
+ * onBatchEnd: Logs include `loss`, and optionally `acc` (if accuracy monitoring
+ *   is enabled).
+ */
+export abstract class Callback {
+  // TODO(michaelterry): This type is a best guess.
+  validationData: Tensor|Tensor[] = null;
+  /** Instance of `keras.models.Model`. Reference of the model being trained. */
+  model: Model = null;
+  /**
+   * Training parameters (eg. verbosity, batch size, number of epochs...).
+   */
+  params: Params;
+
+  setParams(params: Params): void {
+    this.params = params;
+  }
+
+  setModel(model: Model): void {
+    this.model = model;
+  }
+
+  async onEpochBegin(epoch: number, logs?: UnresolvedLogs) {}
+
+  async onEpochEnd(epoch: number, logs?: UnresolvedLogs) {}
+
+  async onBatchBegin(batch: number, logs?: UnresolvedLogs) {}
+
+  async onBatchEnd(batch: number, logs?: UnresolvedLogs) {}
+
+  async onTrainBegin(logs?: UnresolvedLogs) {}
+
+  async onTrainEnd(logs?: UnresolvedLogs) {}
+}
+
+/**
+ * Container abstracting a list of callbacks.
+ */
+export class CallbackList {
+  callbacks: Callback[];
+  queueLength: number;
+
+  // TODO(cais): When the need arises, uncomment the following lines and
+  // implement the queue for time values.
+  // private deltaTBatch: number;
+  // private deltaTsBatchBegin: Array<number>;
+  // private deltaTsBatchEnd: Array<number>;
+
+  /**
+   * Constructor of CallbackList.
+   * @param callbacks Array of `Callback` instances.
+   * @param queueLength Queue length for keeping running statistics over
+   *   callback execution time.
+   */
+  constructor(callbacks?: Callback[], queueLength = 10) {
+    // TODO(cais): Make use of queueLength when implementing the queue for time
+    // values.
+    if (callbacks == null) {
+      callbacks = [];
+    }
+    this.callbacks = callbacks;
+    this.queueLength = queueLength;
+  }
+
+  append(callback: Callback): void {
+    this.callbacks.push(callback);
+  }
+
+  setParams(params: Params): void {
+    for (const callback of this.callbacks) {
+      callback.setParams(params);
+    }
+  }
+
+  setModel(model: Model): void {
+    for (const callback of this.callbacks) {
+      callback.setModel(model);
+    }
+  }
+
+  /**
+   * Called at the start of an epoch.
+   * @param epoch Index of epoch.
+   * @param logs Dictionary of logs.
+   */
+  async onEpochBegin(epoch: number, logs?: UnresolvedLogs) {
+    if (logs == null) {
+      logs = {};
+    }
+    for (const callback of this.callbacks) {
+      await callback.onEpochBegin(epoch, logs);
+    }
+  }
+
+  /**
+   * Called at the end of an epoch.
+   * @param epoch Index of epoch.
+   * @param logs Dictionary of logs.
+   */
+  async onEpochEnd(epoch: number, logs?: UnresolvedLogs) {
+    if (logs == null) {
+      logs = {};
+    }
+    for (const callback of this.callbacks) {
+      await callback.onEpochEnd(epoch, logs);
+    }
+  }
+
+  /**
+   * Called  right before processing a batch.
+   * @param batch Index of batch within the current epoch.
+   * @param logs Dictionary of logs.
+   */
+  async onBatchBegin(batch: number, logs?: UnresolvedLogs) {
+    if (logs == null) {
+      logs = {};
+    }
+    for (const callback of this.callbacks) {
+      await callback.onBatchBegin(batch, logs);
+    }
+  }
+
+  /**
+   * Called at the end of a batch.
+   * @param batch Index of batch within the current epoch.
+   * @param logs Dictionary of logs.
+   */
+  async onBatchEnd(batch: number, logs?: UnresolvedLogs) {
+    if (logs == null) {
+      logs = {};
+    }
+    for (const callback of this.callbacks) {
+      await callback.onBatchEnd(batch, logs);
+    }
+  }
+
+  /**
+   * Called at the beginning of training.
+   * @param logs Dictionary of logs.
+   */
+  async onTrainBegin(logs?: UnresolvedLogs) {
+    if (logs == null) {
+      logs = {};
+    }
+    for (const callback of this.callbacks) {
+      await callback.onTrainBegin(logs);
+    }
+  }
+
+  /**
+   * Called at the end of training.
+   * @param logs Dictionary of logs.
+   */
+  async onTrainEnd(logs?: UnresolvedLogs) {
+    if (logs == null) {
+      logs = {};
+    }
+    for (const callback of this.callbacks) {
+      await callback.onTrainEnd(logs);
+    }
+  }
+}
+
+
+/**
+ * Standardize callbacks or configurations of them to an Array of callbacks.
+ */
+export function standardizeCallbacks(callbacks: Callback|Callback[]|
+                                     CustomCallbackConfig|
+                                     CustomCallbackConfig[]): Callback[] {
+  if (callbacks == null) {
+    return null;
+  }
+  if (callbacks instanceof Callback) {
+    return [callbacks as Callback];
+  }
+  if (Array.isArray(callbacks) && callbacks[0] instanceof Callback) {
+    return callbacks as Callback[];
+  }
+  // Convert custom callback configs to custom callback objects.
+  const callbackConfigs =
+      generic_utils.toList(callbacks) as CustomCallbackConfig[];
+  return callbackConfigs.map(
+      callbackConfig => new CustomCallback(callbackConfig));
+}
+
+export interface CustomCallbackConfig {
+  onTrainBegin?: (logs?: Logs) => Promise<void>;
+  onTrainEnd?: (logs?: Logs) => Promise<void>;
+  onEpochBegin?: (epoch: number, logs?: Logs) => Promise<void>;
+  onEpochEnd?: (epoch: number, logs?: Logs) => Promise<void>;
+  onBatchBegin?: (batch: number, logs?: Logs) => Promise<void>;
+  onBatchEnd?: (batch: number, logs?: Logs) => Promise<void>;
+}
+
+/**
+ * Custom callback for training.
+ */
+export class CustomCallback extends Callback {
+  protected readonly trainBegin: (logs?: Logs) => Promise<void>;
+  protected readonly trainEnd: (logs?: Logs) => Promise<void>;
+  protected readonly epochBegin: (epoch: number, logs?: Logs) => Promise<void>;
+  protected readonly epochEnd: (epoch: number, logs?: Logs) => Promise<void>;
+  protected readonly batchBegin: (batch: number, logs?: Logs) => Promise<void>;
+  protected readonly batchEnd: (batch: number, logs?: Logs) => Promise<void>;
+
+  constructor(config: CustomCallbackConfig) {
+    super();
+    this.trainBegin = config.onTrainBegin;
+    this.trainEnd = config.onTrainEnd;
+    this.epochBegin = config.onEpochBegin;
+    this.epochEnd = config.onEpochEnd;
+    this.batchBegin = config.onBatchBegin;
+    this.batchEnd = config.onBatchEnd;
+  }
+
+  async onEpochBegin(epoch: number, logs?: UnresolvedLogs): Promise<void> {
+    if (this.epochBegin != null) {
+      await resolveScalarsInLogs(logs);
+      await this.epochBegin(epoch, logs as Logs);
+    }
+  }
+
+  async onEpochEnd(epoch: number, logs?: UnresolvedLogs): Promise<void> {
+    if (this.epochEnd != null) {
+      await resolveScalarsInLogs(logs);
+      await this.epochEnd(epoch, logs as Logs);
+    }
+  }
+
+  async onBatchBegin(batch: number, logs?: UnresolvedLogs): Promise<void> {
+    if (this.batchBegin != null) {
+      await resolveScalarsInLogs(logs);
+      await this.batchBegin(batch, logs as Logs);
+    }
+  }
+
+  async onBatchEnd(batch: number, logs?: UnresolvedLogs): Promise<void> {
+    if (this.batchEnd != null) {
+      await resolveScalarsInLogs(logs);
+      await this.batchEnd(batch, logs as Logs);
+    }
+  }
+
+  async onTrainBegin(logs?: UnresolvedLogs): Promise<void> {
+    if (this.trainBegin != null) {
+      await resolveScalarsInLogs(logs);
+      await this.trainBegin(logs as Logs);
+    }
+  }
+
+  async onTrainEnd(logs?: UnresolvedLogs): Promise<void> {
+    if (this.trainEnd != null) {
+      await resolveScalarsInLogs(logs);
+      await this.trainEnd(logs as Logs);
+    }
+  }
+}
+
+
+/**
+ * Callback that accumulates epoch averages of metrics.
+ *
+ * This callback is automatically applied to every Model.
+ */
+export class BaseLogger extends Callback {
+  private seen: number;
+  private totals: UnresolvedLogs;
+
+  constructor() {
+    super();
+  }
+
+  async onEpochBegin(epoch: number, logs?: UnresolvedLogs) {
+    this.seen = 0;
+    this.totals = {};
+  }
+
+  async onBatchEnd(batch: number, logs?: UnresolvedLogs) {
+    if (logs == null) {
+      logs = {};
+    }
+    const batchSize = logs['size'] == null ? 0 : logs['size'] as number;
+    this.seen += batchSize;
+    for (const key in logs) {
+      const value = logs[key];
+      if (typeof value === 'number') {
+        if (!this.totals.hasOwnProperty(key)) {
+          this.totals[key] = 0;
+        }
+        this.totals[key] = this.totals[key] as number + value * batchSize;
+      } else {
+        let oldTotalsToDispose: Scalar;
+        if (key in this.totals) {
+          oldTotalsToDispose = this.totals[key] as Scalar;
+        } else {
+          this.totals[key] = getScalar(0);
+        }
+
+        this.totals[key] = tidy(
+            () => add((this.totals[key] as Scalar),
+                      mul(value, getScalar(batchSize))) as Scalar);
+        if (oldTotalsToDispose != null) {
+          oldTotalsToDispose.dispose();
+        }
+      }
+    }
+  }
+
+  async onEpochEnd(epoch: number, logs?: UnresolvedLogs) {
+    if (logs != null) {
+      for (const key of this.params['metrics'] as string[]) {
+        if (this.totals[key] == null) {
+          continue;
+        }
+        if (typeof this.totals[key] === 'number') {
+          logs[key] = this.totals[key] as number / this.seen;
+        } else {
+          tidy(() => {
+            logs[key] = tfc.mul(
+                            div(getScalar(1), getScalar(this.seen)) as Scalar,
+                            this.totals[key] as Scalar) as Scalar;
+            (this.totals[key] as Tensor).dispose();
+            keep(logs[key] as Scalar);
+          });
+        }
+      }
+    }
+  }
+}
+
+/**
+ * Callback that records events into a `History` object. This callback is
+ * automatically applied to every TF.js Layers model. The `History` object gets
+ * returned by the `fit` method of models.
+ */
+export class History extends Callback {
+  epoch: number[];
+  history: {[key: string]: Array<number|Tensor>};
+
+  async onTrainBegin(logs?: UnresolvedLogs) {
+    this.epoch = [];
+    this.history = {};
+  }
+
+  async onEpochEnd(epoch: number, logs?: UnresolvedLogs) {
+    if (logs == null) {
+      logs = {};
+    }
+    this.epoch.push(epoch);
+    for (const key in logs) {
+      if (this.history[key] == null) {
+        this.history[key] = [];
+      }
+      this.history[key].push(logs[key]);
+    }
+  }
+
+  /**
+   * Await the values of all losses and metrics.
+   */
+  async syncData() {
+    const promises: Array<Promise<Float32Array|Int32Array|Uint8Array>> = [];
+    const keys: string[] = [];
+    const indices: number[] = [];
+    for (const key in this.history) {
+      const valueArray = this.history[key];
+      for (let i = 0; i < valueArray.length; ++i) {
+        if (typeof valueArray[i] !== 'number') {
+          const valueScalar = valueArray[i] as Tensor;
+          promises.push(valueScalar.data());
+          keys.push(key);
+          indices.push(i);
+        }
+      }
+    }
+    const values = await Promise.all(promises);
+    for (let n = 0; n < values.length; ++n) {
+      (this.history[keys[n]][indices[n]] as Tensor).dispose();
+      this.history[keys[n]][indices[n]] = values[n][0];
+    }
+  }
+}
