+++ src/layers/merge.ts
@@ -16,11 +16,13 @@ import * as tfc from '@tensorflow/tfjs-core';
 import {serialization, Tensor, tidy, util} from '@tensorflow/tfjs-core';
 
 import * as K from '../backend/tfjs_backend';
-import {Layer, LayerConfig} from '../engine/topology';
+import {Layer, LayerConfig, SymbolicTensor} from '../engine/topology';
+import {getScalar} from '../engine/unique_id';
 import {NotImplementedError, ValueError} from '../errors';
-import {Kwargs, Shape, SymbolicTensor} from '../types';
+import {Kwargs, Shape} from '../types';
 import * as generic_utils from '../utils/generic_utils';
 import * as mathUtils from '../utils/math_utils';
+import {getExactlyOneShape} from '../utils/types_utils';
 
 /**
  * Generic Merge layer for element-wise merge functions.
@@ -87,7 +89,7 @@ export abstract class Merge extends Layer {
     // Used purely for shape validation.
     if (Array.isArray(inputShape) && !Array.isArray(inputShape[0])) {
       // Make sure that inputShape is an Array of shape.
-      inputShape = [generic_utils.getExactlyOneShape(inputShape)];
+      inputShape = [getExactlyOneShape(inputShape)];
     }
     inputShape = inputShape as Shape[];
     if (inputShape.length < 2) {
@@ -153,7 +155,7 @@ export abstract class Merge extends Layer {
           for (const x of inputs) {
             const xNDim = x.rank;
             if (xNDim == null) {
-              const xShape = K.shape(x);
+              const xShape = x.shape;
               const batchSize = xShape[0];
               const newShape = xShape.slice(1).concat([batchSize]);
               let xTransposed = x.reshape(
@@ -177,7 +179,7 @@ export abstract class Merge extends Layer {
             // If inputs have been transposed, we have to transpose the output
             // too.
             if (yNDim == null) {
-              const yShape = K.shape(y);
+              const yShape = y.shape;
               const yNDim = yShape.length;
               const batchSize = yShape[yNDim - 1];
               const newShape =
@@ -441,7 +443,7 @@ export class Average extends Merge {
       for (const input of inputs) {
         output = tfc.add(output, input);
       }
-      return K.scalarTimesArray(K.getScalar(1 / inputs.length), output);
+      return tfc.mul(getScalar(1 / inputs.length), output);
     });
   }
 }
