+++ src/layers/normalization_test.ts
@@ -16,7 +16,7 @@
 import {onesLike, Tensor, tensor1d, tensor2d, tensor3d, tensor4d, train, zeros, zerosLike} from '@tensorflow/tfjs-core';
 
 import * as tfl from '../index';
-import {SymbolicTensor} from '../types';
+import {SymbolicTensor} from '../engine/topology';
 import {describeMathCPU, describeMathCPUAndGPU, expectTensorsClose} from '../utils/test_utils';
 import {batchNormalization, normalizeBatchInTraining} from './normalization';
