+++ src/layers/padding_test.ts
@@ -17,8 +17,8 @@ import * as tfc from '@tensorflow/tfjs-core';
 import {Tensor, tensor3d} from '@tensorflow/tfjs-core';
 
 import {DataFormat} from '../common';
+import {SymbolicTensor} from '../engine/topology';
 import * as tfl from '../index';
-import {SymbolicTensor} from '../types';
 import {convertPythonicToTs, convertTsToPythonic} from '../utils/serialization_utils';
 import {describeMathCPU, describeMathCPUAndGPU, expectTensorsClose} from '../utils/test_utils';
