+++ src/types.ts
@@ -11,81 +11,17 @@
 /** Defines allowable data types for tensors. */
 
 // tslint:disable:max-line-length
-import {DataType, doc, Scalar, Tensor} from '@tensorflow/tfjs-core';
+import {Scalar, Tensor} from '@tensorflow/tfjs-core';
 
-import {getScopedTensorName, getUniqueTensorName} from './common';
-import {Layer} from './engine/topology';
+// import {Layer} from './engine/topology';
 // tslint:enable:max-line-length
 
 /** @docalias number[] */
 export type Shape = number[];
 
-/**
- * An ID to track `SymbolicTensor`s and derived classes.
- * Required in different places in engine/topology.ts to identify unique
- * tensors.
- */
-let _nextUniqueTensorId = 0;
-
-export function getNextUniqueTensorId(): number {
-  return _nextUniqueTensorId++;
-}
-
-/**
- * `SymbolicTensor` is a placeholder for a Tensor without any concrete value.
- *
- * They are most often encountered when building a graph of `Layer`s for a
- * a `Model` and the input data's shape, but not values are known.
- */
-@doc({heading: 'Models', 'subheading': 'Classes'})
-export class SymbolicTensor {
-  /* A unique ID for the tensor to be able to differentiate tensors. */
-  readonly id: number;
-  // The fully scoped name of this Variable, including a unique suffix if needed
-  readonly name: string;
-  // The originally requested fully scoped name of this Variable, not including
-  // any unique suffix.  This may be needed when restoring weights because this
-  // original name is used as a key.
-  readonly originalName?: string;
-  /**
-   * Rank/dimensionality of the tensor.
-   */
-  readonly rank: number;
-  /**
-   * Replacement for _keras_history.
-   */
-  nodeIndex: number;
-  /**
-   * Replacement for _keras_history.
-   */
-  tensorIndex: number;
-
-  /**
-   *
-   * @param dtype
-   * @param shape
-   * @param sourceLayer The Layer that produced this symbolic tensor.
-   * @param inputs The inputs passed to sourceLayer's __call__() method.
-   * @param nodeIndex
-   * @param tensorIndex
-   * @param callArgs The keyword arguments passed to the __call__() method.
-   * @param name
-   * @param outputTensorIndex The index of this tensor in the list of outputs
-   *   returned by apply().
-   */
-  constructor(
-      readonly dtype: DataType, readonly shape: Shape,
-      public sourceLayer: Layer, readonly inputs: SymbolicTensor[],
-      readonly callArgs: Kwargs, name?: string,
-      readonly outputTensorIndex?: number) {
-    this.id = getNextUniqueTensorId();
-    if (name != null) {
-      this.originalName = getScopedTensorName(name);
-      this.name = getUniqueTensorName(this.originalName);
-    }
-    this.rank = shape.length;
-  }
-}
+export type HasShape = {
+  shape: Shape;
+};
 
 /**
  * Type for loss a metric function.
